+++ src_docs/components/date-range.component.jsx
@@ -12,7 +12,10 @@ export default class DateRangeView extends React.PureComponent {
       <>
         <div style={{ padding: '20px' }}>
           <h4>Absolute period (default)</h4>
-          <DateRange id="example1" />
+          <DateRange
+            id="example1"
+            absoluteRange={{ dateFormat: 'DD/MM/YY' }}
+          />
           <h4>Absolute or relative</h4>
           <DateRange id="example2" enabled={{ absolute: true, relative: true }} />
           <h4>Absolute, relative or period</h4>